requirements.txt
flask~=3.1.0
pymongo~=4.12.1
python-dotenv
numpy~=2.2.5
opencv-contrib-python
opencv-python~=4.11.0.86
openpyxl
pandas~=2.2.3
pillow~=11.2.1
pyttsx3~=2.98
flask-cors~=5.0.1